Move files to web/chip8/
